# Don't overwrite Base.vect for numbers with a different implementation