Fixes https://github.com/JuliaOpt/Convex.jl/issues/241