@iamed2 2018-11-02
Double-checked that this compiles out the `any` call?