@iamed2 2018-11-02
Should we do this with `vcat` as well?