@ararslan 2018-11-02
> Double-checked that this compiles out the any call?

It does not, unfortunately, and there is a performance impact on the Base method with this change.

After loading Convex:
```
julia> @benchmark Base.vect($x...)
BenchmarkTools.Trial: 
  memory estimate:  102.72 KiB
  allocs estimate:  375
  --------------
  minimum time:     36.807 μs (0.00% GC)
  median time:      38.912 μs (0.00% GC)
  mean time:        45.703 μs (12.67% GC)
  maximum time:     33.698 ms (99.77% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
(here `x` is a 74-element `Vector{Any}` consisting of `Int`s, `Float32`s, and `Complex{Int}`s.)

Without loading Convex, same `x`:
```
julia> @benchmark Base.vect($x...)
BenchmarkTools.Trial: 
  memory estimate:  55.84 KiB
  allocs estimate:  148
  --------------
  minimum time:     22.036 μs (0.00% GC)
  median time:      23.726 μs (0.00% GC)
  mean time:        28.634 μs (13.89% GC)
  maximum time:     25.832 ms (99.89% GC)
  --------------
  samples:          10000
  evals/sample:     1
```