@ararslan 2018-11-02
Okay, I've reinstated the old code but corrected the implementation.