@iamed2 2018-11-02
Hmm what if we changed it to use `invoke` unconditionally?