@ararslan 2018-11-02
I checked that earlier and there is still a performance impact, unfortunately. Smaller than with the `any` branch at least, but still slower than without loading Convex. With the current state of the PR, the performance is unchanged after loading Convex (since the code does exactly the same thing).